+++ src/KsWare.Presentation.ViewFramework.Common.Tests/DesignTime/PropertiesTests.cs
@@ -1,0 +1,21 @@
+﻿
+using NUnit.Framework;
+
+namespace KsWare.Presentation.ViewFramework.Tests.DesignTime {
+
+	[TestFixture]
+	public class PropertiesTests {
+
+		[Test]
+		public void Test() {
+			//TestCop		
+			Assert.Fail("WriteMe");
+		}
+
+		
+
+	}
+
+}
+
+